fix trainer screen layout
